telegram/pyproject.toml
[project]
name = "perekup"
version = "0.1.0"
description = ""
authors = [{ name = "alex" }]
requires-python = "~=3.12"
readme = "README.md"
dependencies = [
    "asyncpg>=0.30.0",
    "pydantic>=2.11.4",
    "pydantic-settings>=2.9.1",
    "python-dotenv>=1.1.0",
    "sqlalchemy[asyncio]>=2.0.40",
    "telethon>=1.40.0",
    "tgcrypto>=1.2.5",
]
